% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{option_price_call_black_scholes}
\alias{option_price_call_black_scholes}
\title{B-S Option Pricing Model}
\usage{
option_price_call_black_scholes(S, K, r, sigma, time)
}
\arguments{
\item{S}{underlying asset price}

\item{K}{exercise price}

\item{r}{risk free interest rate}

\item{sigma}{volatility of assets(year)}

\item{time}{term of contract}
}
\value{
the price of one European call option
}
\description{
use B-S option pricing model to compute the price of European call option
}
\examples{
\dontrun{
option_price_call_black_scholes(4.019, 3.3, 0.022685, 0.4199, 0.05479)
}
}
