\name{StatComp22095-package}
\alias{StatComp22095-package}
\docType{package}
\title{
  This package provides several functions to compute VAR/ES/volatility/options price
}
\description{
  Compute the dynamic volatility of a stock by SMA/GMA/EWMA; use garch model to compute the VAR and ES of one stock; use B-S option pricing model to compute the price of European option
}
\author{
Bowen Tong

Maintainer: Bowen Tong <1091755418@qq.com>
}
\keyword{ package }
