% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR22095.R
\name{VAR_and_ES}
\alias{VAR_and_ES}
\title{Compute VAR and ES}
\usage{
VAR_and_ES(data)
}
\arguments{
\item{data}{the close price of the stock}
}
\value{
VAR and ES for this stock and some figures
}
\description{
use garch model to compute the VAR and ES of one stock
}
\examples{
\dontrun{
library(StatComp22095)
data(data)
VAR_and_ES(data)
}
}
