% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR22095.R
\name{volatility}
\alias{volatility}
\title{Compute volatility}
\usage{
volatility(data, SMA = TRUE, GMA = TRUE, EWMA = TRUE)
}
\arguments{
\item{data}{the close price of the stock}

\item{SMA}{return the SMA result or not}

\item{GMA}{return the GMA result or not}

\item{EWMA}{return the EWMA result or not}
}
\value{
the dynamic volatility and some figures
}
\description{
compute the dynamic volatility of a stock by SMA/GMA/EWMA
}
\examples{
\dontrun{
library(StatComp22095)
data(data)
volatility(data, EWMA = TRUE, GMA = FALSE, SMA = FALSE)
}
}
